requirements.txt
# Django & REST Framework
Django==4.2.11
djangorestframework==3.14.0
whitenoise==6.6.0

# Channels & WebSocket & ASGI
channels==4.0.0
channels-redis==4.1.0
daphne==4.0.0

# Database
psycopg2-binary==2.9.9

# Redis & Celery
redis==5.0.1
celery==5.3.4

# Computer Vision & Face Recognition (GPU-enabled)
numpy==1.24.3
opencv-python-headless==4.8.1.78
insightface==0.7.3
onnxruntime-gpu==1.17.1

# Image Processing
Pillow==10.1.0

# Additional dependencies
python-dateutil==2.8.2
pytz==2023.3
